Use row timestamp for ready-to-merge reconciliation
